:recycle: Rename all occurrences of AccessGrant#ownerId with AccessGrant#creatorId.
For access restriction, ownership is taken from the dataset, not the access grant.
In the future it might make sense to create a dataset_stewardships table that contains the department/division/branch/unit information similar to the user_group_memberships table and use that to control dataset access.

Currently the dataset owner is set on record creation from the dataset owner, and can't be changed after creation. This is clearly not going to work in the long run.