:art: Break nested queries out into name variables.
Even if I can't use common table expressions here, as this is a sub-query,
It makes it much easier to read than the 3 layer nested query.

At some future point, if speed becomes and issue, it will make sense to write the entire controller logic, including policy scoping as a single optimized query.