:beetle: Fix csv download failure error writing.
Writing to the stream fails because it expects the content to be formatted as a CSV.
We can bypass this limitation by writing directly to the output stream.

A better UI would be to return a 500 on failure, instead of writing to the file.
With that said, due to the slowness of the api refresh operation, this means file download won't start for 30 seconds, and which is also bad UI.