:bug: Fix policy scope usage in AccessRequestsController.
1. Fix some type continuity issues. Now the policy types will complain
unless you set
```ts
static policyScope(user: User): FindOptions<Attributes<SomeSpecificModel>> {}
```

I'd prefer if the types were generated magically, but can't get it to work.
At least this way they still fail correctly.

2. Encountered an issue with using scopes inside of other scopes, and decided
inlining the code was the least painful way to fix it.

The alternatives are:
a. Add a method to the class that returns the scope, and use it in both places.
   Best if the scope is used in many places.
b. `AccessRequest.scope({ method: ["withDatasetOwnerId", user.id] })._scope`.
   this is a hack loading info from a private method, and is not recommended.
   But it _would_ work, so good to be aware of it.
c. Increase the complexity of the scope generator to handle returning a scope.
   This vastly increases the complexity of the code, and types, so probably not a good choice
   long term.